refactor: split pages/index.js into 8 focused section components under _sections/, remove unused ScrollHint
